+++ scifit/tests/test_kinetics.py
@@ -37,12 +37,24 @@ class GenericASMKineticTest:
     def test_reactant_references(self):
         references = self.solver.references
 
+    def test_model_formula(self):
+        formula = self.solver.model_formulas()
+
     def test_conversion_ratio(self):
         self.solver.solve(t=self.t)
         ratio = self.solver.convertion_ratio()
 
-    def test_model_formula(self):
-        formula = self.solver.model_formulas()
+    def test_first_derivative(self):
+        self.solver.solve(t=self.t)
+        dxdt = self.solver.derivative(derivative_order=1)
+
+    def test_second_derivative(self):
+        self.solver.solve(t=self.t)
+        d2xdt2 = self.solver.derivative(derivative_order=2)
+
+    def test_selectivities(self):
+        self.solver.solve(t=self.t)
+        selectivities = self.solver.selectivities()
 
     def test_plot_solve(self):
         name = self.__class__.__name__
@@ -55,21 +67,48 @@ class GenericASMKineticTest:
         name = self.__class__.__name__
         self.solver.solve(t=self.t)
         axe = self.solver.plot_solve_ratio()
-        axe.figure.savefig("{}/{}_solve_ratio.{}".format(self.media_path, name, self.format))
+        axe.figure.savefig(
+            "{}/{}_solve_ratio.{}".format(self.media_path, name, self.format)
+        )
+        plt.close(axe.figure)
+
+    def test_plot_first_derivative(self):
+        name = self.__class__.__name__
+        self.solver.solve(t=self.t)
+        axe = self.solver.plot_first_derivative()
+        axe.figure.savefig("{}/{}_dxdt.{}".format(self.media_path, name, self.format))
+        plt.close(axe.figure)
+
+    def test_plot_second_derivative(self):
+        name = self.__class__.__name__
+        self.solver.solve(t=self.t)
+        axe = self.solver.plot_second_derivative()
+        axe.figure.savefig("{}/{}_d2xdt2.{}".format(self.media_path, name, self.format))
+        plt.close(axe.figure)
+
+    def test_plot_selectivities(self):
+        name = self.__class__.__name__
+        self.solver.solve(t=self.t)
+        axe = self.solver.plot_selectivities()
+        axe.figure.savefig("{}/{}_selectivities.{}".format(self.media_path, name, self.format))
         plt.close(axe.figure)
 
     def test_plot_quotients(self):
         name = self.__class__.__name__
         self.solver.solve(t=self.t)
         axe = self.solver.plot_quotients()
-        axe.figure.savefig("{}/{}_quotients.{}".format(self.media_path, name, self.format))
+        axe.figure.savefig(
+            "{}/{}_quotients.{}".format(self.media_path, name, self.format)
+        )
         plt.close(axe.figure)
 
 
 class SimpleKinetic01(GenericASMKineticTest):
-    nur = np.array([
-        [-1, 1],
-    ])
+    nur = np.array(
+        [
+            [-1, 1],
+        ]
+    )
     nup = None
     x0 = np.array([2e-3, 0.0])
     k0 = np.array([1e-2])
@@ -89,9 +128,11 @@ class SimpleKinetic01K0L2(SimpleKinetic01, TestCase):
 
 
 class SimpleKinetic02(GenericASMKineticTest):
-    nur = np.array([
-        [-1, -1, 1],
-    ])
+    nur = np.array(
+        [
+            [-1, -1, 1],
+        ]
+    )
     nup = None
     x0 = np.array([2e-3, 5e-3, 0.0])
     k0 = np.array([1e-0])
@@ -103,17 +144,19 @@ class SimpleKinetic02K0L0(SimpleKinetic02, TestCase):
 
 
 class SimpleKinetic02K0L1(SimpleKinetic02, TestCase):
-    k0 = np.array([1e+1])
+    k0 = np.array([1e1])
 
 
 class SimpleKinetic02K0L2(SimpleKinetic02, TestCase):
-    k0 = np.array([1e+2])
+    k0 = np.array([1e2])
 
 
 class SimpleKinetic03(GenericASMKineticTest):
-    nur = np.array([
-        [-1, 1],
-    ])
+    nur = np.array(
+        [
+            [-1, 1],
+        ]
+    )
     nup = None
     mode = "indirect"
     x0 = np.array([0.0, 2e-3])
@@ -131,13 +174,15 @@ class SimpleKinetic03K0L1(SimpleKinetic03, TestCase):
 
 
 class SimpleKinetic03K0L2(SimpleKinetic03, TestCase):
-    k0inv = np.array([1e+0])
+    k0inv = np.array([1e0])
 
 
 class SimpleKinetic04(GenericASMKineticTest):
-    nur = np.array([
-        [-1, 1],
-    ])
+    nur = np.array(
+        [
+            [-1, 1],
+        ]
+    )
     nup = None
     mode = "equilibrium"
     x0 = np.array([3e-3, 1e-3])
@@ -159,12 +204,16 @@ class SimpleKinetic04K0L2(SimpleKinetic04, TestCase):
 
 
 class SimpleKinetic05(GenericASMKineticTest):
-    nur = np.array([
-        [-1, -1],
-    ])
-    nup = np.array([
-        [0,   2],
-    ])
+    nur = np.array(
+        [
+            [-1, -1],
+        ]
+    )
+    nup = np.array(
+        [
+            [0, 2],
+        ]
+    )
     x0 = np.array([5e-3, 2e-3])
     k0 = np.array([1e-1])
     t = np.linspace(0.0, 500.0, 1000)
@@ -175,18 +224,20 @@ class SimpleKinetic05K0L0(SimpleKinetic05, TestCase):
 
 
 class SimpleKinetic05K0L1(SimpleKinetic05, TestCase):
-    k0 = np.array([1e+0])
+    k0 = np.array([1e0])
 
 
 class SimpleKinetic05K0L2(SimpleKinetic05, TestCase):
-    k0 = np.array([1e+1])
+    k0 = np.array([1e1])
 
 
 class MultipleKinetics01(GenericASMKineticTest, TestCase):
-    nur = np.array([
-        [-1, 1, 0],
-        [0, -1, 1],
-    ])
+    nur = np.array(
+        [
+            [-1, 1, 0],
+            [0, -1, 1],
+        ]
+    )
     nup = None
     x0 = np.array([2e-3, 0.0, 0.0])
     k0 = np.array([1e-2, 1e-3])
@@ -194,11 +245,13 @@ class MultipleKinetics01(GenericASMKineticTest, TestCase):
 
 
 class MultipleKinetics02(GenericASMKineticTest, TestCase):
-    nur = np.array([
-        [-1, 1, 0, 0],
-        [0, -1, 1, 0],
-        [0, 0, -1, 1],
-    ])
+    nur = np.array(
+        [
+            [-1, 1, 0, 0],
+            [0, -1, 1, 0],
+            [0, 0, -1, 1],
+        ]
+    )
     nup = None
     x0 = np.array([2e-3, 0.0, 0.0, 0.0])
     k0 = np.array([1e-2, 1e-3, 1e-1])
@@ -206,18 +259,13 @@ class MultipleKinetics02(GenericASMKineticTest, TestCase):
 
 
 class MultipleKinetics03(GenericASMKineticTest, TestCase):
-    nur = np.array([
-        [-1, 1, 0, 0, 0],
-        [0, -1, 1, 0, 0],
-        [0, -1, 0, 1, 0],
-        [0, 0, -1, -1, 1]
-    ])
+    nur = np.array(
+        [[-1, 1, 0, 0, 0], [0, -1, 1, 0, 0], [0, -1, 0, 1, 0], [0, 0, -1, -1, 1]]
+    )
     nup = None
     x0 = np.array([2e-3, 0.0, 0.0, 0.0, 0.0])
     k0 = np.array([1e-2, 3e-2, 1e-1, 1e-2])
     t = np.linspace(0.0, 500.0, 500)
-
-
 
 
 # class SimpleKinetic07(GenericASMKineticTest, TestCase):